pyproject.toml
[project]
name = "mini-hackathon-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.0"
dependencies = [
    "accelerate>=1.13.0",
    "anthropic>=0.86.0",
    "diffusers",
    "dotenv>=0.9.9",
    "flask>=3.1.3",
    "google-genai>=1.68.0",
    "gradio>=5.0.0",
    "json-repair>=0.58.6",
    "matplotlib>=3.10.8",
    "openai>=2.29.0",
    "pillow>=12.1.1",
    "python-pptx>=1.0.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.3.0",
]

[tool.uv.sources]
diffusers = { git = "https://github.com/huggingface/diffusers" }
